refactor components to common file; add features in galactic map
